Relocate misc functions to separate files
