+++ lib/src/solid/chk_exists_and_has_acl.dart
@@ -36,6 +36,7 @@ import 'package:solidpod/src/solid/api/common_permission.dart';
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants/common.dart';
 import 'package:solidpod/src/solid/solid_func_call_status.dart';
+import 'package:solidpod/src/solid/utils/get_url_helper.dart';
 import 'package:solidpod/src/solid/utils/misc.dart';
 
 /// Check [fileName] exists and has the associated ACL file. Requires user
