+++ lib/src/solid/read_permission.dart
@@ -30,6 +30,7 @@ library;
 
 import 'dart:core';
 
+import 'package:solidpod/src/solid/utils/get_url_helper.dart';
 import 'package:solidpod/src/solid/utils/misc.dart';
 import 'package:solidpod/src/solid/utils/permission.dart';
