+++ lib/src/solid/utils/misc.dart
@@ -30,11 +30,8 @@
 
 library;
 
-import 'dart:convert';
-
 import 'package:flutter/foundation.dart' show debugPrint;
 
-import 'package:crypto/crypto.dart';
 import 'package:encrypter_plus/encrypter_plus.dart';
 import 'package:fast_rsa/fast_rsa.dart' show KeyPair;
 import 'package:intl/intl.dart';
@@ -51,7 +48,10 @@ import 'package:solidpod/src/solid/constants/web_acl.dart';
 import 'package:solidpod/src/solid/revoke_permission_to_recipients.dart';
 import 'package:solidpod/src/solid/utils/app_info.dart';
 import 'package:solidpod/src/solid/utils/authdata_manager.dart';
+import 'package:solidpod/src/solid/utils/data_encryption.dart';
 import 'package:solidpod/src/solid/utils/exceptions.dart';
+import 'package:solidpod/src/solid/utils/get_url_helper.dart';
+import 'package:solidpod/src/solid/utils/init_pod_gen_resources.dart';
 import 'package:solidpod/src/solid/utils/key_manager.dart';
 import 'package:solidpod/src/solid/utils/permission.dart';
 import 'package:solidpod/src/solid/utils/rdf.dart';
@@ -84,19 +84,6 @@ Future<void> writeToSecureStorage(String key, String value) async {
     value: value,
   );
 }
-
-/// Encrypt data using AES with the specified key
-String encryptData(String data, Key key, IV iv, {AESMode mode = AESMode.sic}) =>
-    Encrypter(AES(key, mode: mode)).encrypt(data, iv: iv).base64;
-
-/// Decrypt a ciphertext value
-String decryptData(
-  String encData,
-  Key key,
-  IV iv, {
-  AESMode mode = AESMode.sic,
-}) =>
-    Encrypter(AES(key, mode: mode)).decrypt(Encrypted.from64(encData), iv: iv);
 
 /// Load and parse a private TTL file from POD
 // Future<Map<String, dynamic>> loadPrvTTL(String fileUrl) async {
@@ -137,131 +124,6 @@ String decryptData(
 //     throw Exception('Failed to read encryption key file: $e');
 //   }
 // }
-
-/// Generates a public key block from a given key content.
-String genPubKeyStr(String pubKeyContent) =>
-    '''-----BEGIN RSA PUBLIC KEY-----\n$pubKeyContent\n-----END RSA PUBLIC KEY-----''';
-
-/// Get unique bit of the webId
-String getUniqueStrWebId(String webId) {
-  var uniqueStr = webId.replaceAll('https://', '');
-  uniqueStr = uniqueStr.replaceAll('http://', '');
-  uniqueStr = uniqueStr.replaceAll('/$profCard', '');
-  uniqueStr = uniqueStr.replaceAll('/', '-');
-
-  return uniqueStr;
-}
-
-/// Generate unique ID for the resource URL
-/// In order for the ID to be unique across all webIDs and no one can
-/// reverse engineer the resource being shared we also use receiver webId
-/// in the ID generation process
-String getUniqueIdResUrl(String resourceUrl, String receiverWebId) {
-  return sha256.convert(utf8.encode(resourceUrl + receiverWebId)).toString();
-}
-
-/// From a given resource path [resourcePath] create its URL
-/// [resourcePath] should be relative to the POD, e.g., `myapp/data/abc.ttl`.
-/// [isFile] should be true if the resource is a file, otherwise false if it
-/// is a directory.
-/// [webId] (optionally) specifies the resource's POD (could be external).
-/// returns the full resource URL
-
-Future<String> _getResourceUrl(
-  String resourcePath, {
-  required bool isFile,
-  String? webId,
-}) async {
-  final wId = webId ?? await AuthDataManager.getWebId();
-
-  assert(wId != null);
-  assert(wId!.contains(profCard));
-
-  final resourceUrl = wId!.replaceAll(profCard, resourcePath);
-
-  if (!isFile && !resourceUrl.endsWith('/')) {
-    return '$resourceUrl/';
-  }
-
-  return resourceUrl;
-}
-
-/// Create the URL for a file
-Future<String> getFileUrl(String filePath, {String? webId}) async =>
-    await _getResourceUrl(filePath, isFile: true, webId: webId);
-
-/// Create the URL for a directory (container)
-Future<String> getDirUrl(String dirPath, {String? webId}) async =>
-    await _getResourceUrl(dirPath, isFile: false, webId: webId);
-
-/// Get resource Url from a filename, with different options for how
-/// the filename is provided. If [isExternalRes] or [isFileUrl] is
-/// set to true, the filename is already the resource url and is
-/// returned unchanged.
-///
-/// Parameters:
-/// - [fileName] - name of file of interest.
-/// - [isFile] - flag describing whether the resources is a file or not.
-/// If false the [fileName] is a directory. (Default: true).
-/// - [isFilePath] - flag describing whether [fileName] includes the
-/// directory data path. (Default: false).
-/// - [isFileUrl] - flag describing whether [fileName] is the url of the
-/// file. (Default: false).
-/// - [isExternalRes] - flag describing whether the file is an external
-/// file shared with the user. In which case the [fileName] should be
-/// the full URL of the file. (Default: false).
-
-Future<String> filenameToResourceUrl({
-  required String fileName,
-  bool isFile = true,
-  bool isFileUrl = false,
-  bool isExternalRes = false,
-}) async {
-  final String resourceUrl;
-  final String filePath;
-
-  // Note: external resources always specified by url
-  if (isExternalRes) {
-    isFileUrl = true;
-  }
-
-  // If not already a url, get url
-  if (!isExternalRes && !isFileUrl) {
-    // Get the file path
-    // Ensure path uses correct path separators and
-    // has app data dir prepended.
-    filePath = await normalizeFilePath(fileName, null);
-
-    // Get the url of the resource by prepending the
-    // user = owner webID (excluding profCard)
-    if (isFile) {
-      resourceUrl = await getFileUrl(filePath);
-    } else {
-      resourceUrl = await getDirUrl(filePath);
-    }
-  } else {
-    // Use fileName, fileName already the resource url
-    resourceUrl = fileName;
-  }
-
-  return resourceUrl;
-}
-
-/// Derive external POD inherited parent directory url from the resource url
-/// and parent directory path
-String getExtDirUrl(String resourceUrl, String dirPath) {
-  int pathIndex = resourceUrl.indexOf(dirPath);
-
-  String dirUrl = '';
-  if (pathIndex != -1) {
-    dirUrl = resourceUrl.substring(0, pathIndex + dirPath.length);
-    if (!dirUrl.endsWith('/')) {
-      dirUrl += '/';
-    }
-  }
-
-  return dirUrl;
-}
 
 /// Encrypt a given data string and format to TTL
 Future<String> getEncTTLStr(
@@ -370,76 +232,6 @@ Future<void> createDir(String dirUrl) async {
 /// returns true if successful
 
 Future<bool> deleteLogIn() async => AuthDataManager.removeAuthData();
-
-/// Generates a list of default folder paths for a given application.
-///
-/// This function takes the name of an application as input and returns a list of strings.
-/// Each string in the list represents a path to a default folder for the application.
-
-Future<List<String>> generateDefaultFolders() async {
-  final dataDirLoc = [appDirName, dataDir].join('/');
-  final sharingDirLoc = [appDirName, sharingDir].join('/');
-  final sharedDirLoc = [appDirName, sharedDir].join('/');
-  final encDirLoc = [appDirName, encDir].join('/');
-  final logDirLoc = [appDirName, logsDir].join('/');
-
-  final folders = [
-    appDirName,
-    sharingDirLoc,
-    sharedDirLoc,
-    dataDirLoc,
-    encDirLoc,
-    logDirLoc,
-  ];
-  return folders;
-}
-
-/// Generates a list of custom folder paths for a given application.
-///
-/// This function takes the custom folder list given by the developer and
-/// creates paths for those folders. Each custom folder will be created
-/// inside the "data" folder. Following are few examples.
-///   - Custom folder 'myDir1' will be created as '/data/myDir1'
-///   - Custom folder 'myDir1/myDir3' will be created as '/data/myDir1/myDir3'
-///   - Custom folder 'data' will be created as '/data/data'
-
-List<String> generateCustomFolders(List customFolderPaths) {
-  List<String> customFolderList = [];
-  if (customFolderPaths.isNotEmpty) {
-    for (String folderPath in customFolderPaths) {
-      // Remove any leading '/' characters
-      folderPath = folderPath.replaceFirst(RegExp(r'^/+'), '');
-      customFolderList.add([appDirName, dataDir, folderPath].join('/'));
-    }
-  }
-  return customFolderList;
-}
-
-/// Generates a list of default folder paths for a given application.
-///
-/// This function takes the name of an application as input and returns a list of strings.
-/// Each string in the list represents a path to a default folder for the application.
-
-Future<Map<dynamic, dynamic>> generateDefaultFiles() async {
-  final sharingDirLoc = [appDirName, sharingDir].join('/');
-  final sharedDirLoc = [appDirName, sharedDir].join('/');
-  final encDirLoc = [appDirName, encDir].join('/');
-  final logDirLoc = [appDirName, logsDir].join('/');
-
-  final files = {
-    sharingDirLoc: [
-      pubKeyFile,
-      '$pubKeyFile.acl',
-    ],
-    logDirLoc: [
-      permLogFile,
-      '$permLogFile.acl',
-    ],
-    sharedDirLoc: ['.acl'],
-    encDirLoc: [encKeyFile, indKeyFile],
-  };
-  return files;
-}
 
 /// Set directory name for the app for storing the POD data
 ///
